requirements.txt
aniso8601==8.0.0
betacode-flask-restful-swagger-2==0.37
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.25.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
